+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Minor internal refactors & additional tests
 -   Now assuming latest token version if the claim is missing from the header
+-   Make `verifySession` send the appropriate response instead of throwing (and sending 500) if session validation fails in koa, hapi and loopback (already happening in other frameworks).
 
 ## [14.1.1] - 2023-05-24